requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.5.1+cu124
torchvision==0.20.1+cu124
torchaudio==2.5.1+cu124
syncedlyrics
qdrant-client
sentence-transformers
mutagen
tqdm
numpy
librosa
scikit-learn
transformers
openai
pydantic-ai
laion-clap
torchvision
fastapi
uvicorn[standard]
httpx
pytest-asyncio
